+++ app/assets/javascripts/camaleon_cms/admin/tinymce/langs/nl.js
@@ -121,7 +121,7 @@ tinymce.addI18n('nl',{
 "Insert link": "Hyperlink invoegen",
 "New window": "Nieuw venster",
 "None": "Geen",
-"The URL you entered seems to be an external link. Do you want to add the required http:\/\/ prefix?": "De ingegeven URL verwijst naar een extern adres. Wil je er \"http:\/\/\" aan toevoegen?",
+"The URL you entered seems to be an external link. Do you want to add the required https:\/\/ prefix?": "De ingegeven URL verwijst naar een extern adres. Wil je er \"https:\/\/\" aan toevoegen?",
 "Target": "Doel",
 "The URL you entered seems to be an email address. Do you want to add the required mailto: prefix?": "De ingegeven URL lijkt op een e-mailadres. Wil je er \"mailto:\" aan toevoegen?",
 "Insert\/edit link": "Hyperlink invoegen\/bewerken",