+++ app/decorators/camaleon_cms/category_decorator.rb
@@ -21,7 +21,7 @@ module CamaleonCms
 
     # return all children categories for the current category (active_record) filtered by permissions + hidden posts + roles + etc...
     # in return object, you can add custom where's or pagination like here:
-    # http://edgeguides.rubyonrails.org/active_record_querying.html
+    # https://edgeguides.rubyonrails.org/active_record_querying.html
     def the_categories
       object.children
     end