+++ app/helpers/camaleon_cms/theme_helper.rb
@@ -26,7 +26,7 @@ module CamaleonCms
     # asset: (String) asset name
     # theme_name: (optional) theme name, default (current theme caller to this function)
     # sample:
-    #   theme_asset_url("css/main.css") => return: http://myhost.com/assets/themes/my_theme/assets/css/main-54505620f.css
+    #   theme_asset_url("css/main.css") => return: https://myhost.com/assets/themes/my_theme/assets/css/main-54505620f.css
     def theme_asset_url(asset, theme_name = nil)
       p = theme_asset_path(asset, theme_name)
       begin